extract distribute-generic patterns for canonicalization
