+++ lib/Dialect/Secret/IR/BUILD
@@ -15,6 +15,7 @@ cc_library(
     ],
     deps = [
         ":SecretOps",
+        ":SecretPatterns",
         "@heir//include/Dialect/Secret/IR:dialect_inc_gen",
         "@heir//include/Dialect/Secret/IR:ops_inc_gen",
         "@heir//include/Dialect/Secret/IR:types_inc_gen",
@@ -22,6 +23,29 @@ cc_library(
         "@llvm-project//mlir:ControlFlowInterfaces",
         "@llvm-project//mlir:IR",
         "@llvm-project//mlir:InferTypeOpInterface",
+    ],
+)
+
+cc_library(
+    name = "SecretPatterns",
+    srcs = [
+        "SecretPatterns.cpp",
+    ],
+    hdrs = [
+        "@heir//include/Dialect/Secret/IR:SecretDialect.h",
+        "@heir//include/Dialect/Secret/IR:SecretOps.h",
+        "@heir//include/Dialect/Secret/IR:SecretPatterns.h",
+        "@heir//include/Dialect/Secret/IR:SecretTypes.h",
+    ],
+    deps = [
+        "@heir//include/Dialect/Secret/IR:dialect_inc_gen",
+        "@heir//include/Dialect/Secret/IR:ops_inc_gen",
+        "@heir//include/Dialect/Secret/IR:types_inc_gen",
+        "@llvm-project//llvm:Support",
+        "@llvm-project//mlir:ControlFlowInterfaces",
+        "@llvm-project//mlir:IR",
+        "@llvm-project//mlir:InferTypeOpInterface",
+        "@llvm-project//mlir:Support",
     ],
 )
 
@@ -36,6 +60,7 @@ cc_library(
         "@heir//include/Dialect/Secret/IR:SecretTypes.h",
     ],
     deps = [
+        ":SecretPatterns",
         "@heir//include/Dialect/Secret/IR:dialect_inc_gen",
         "@heir//include/Dialect/Secret/IR:ops_inc_gen",
         "@heir//include/Dialect/Secret/IR:types_inc_gen",
